refactor: extracted CRD generation info classes to top level
